sp500-events/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.2
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
icalendar==6.0.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
resend==2.4.0
aiofiles==24.1.0
